chore: Split status-indicator interfaces
